Extract Mantine & AppShell component to Storybook root preview
